requirements/base.txt
Django==2.2.5
django-environ==0.4.5
psycopg2==2.8.4
gunicorn==20.0.4
whitenoise==5.0.1